pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wirlwind-telemetry"
version = "0.1.0"
description = "Real-time network device telemetry dashboard"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Scott Peterman"}
]
dependencies = [
    "PyQt6>=6.5",
    "PyQt6-WebEngine>=6.5",
    "paramiko>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.0",
]

[project.scripts]
wirlwind-telemetry = "wirlwind_telemetry.__main__:main"

[tool.setuptools.packages.find]
include = ["wirlwind_telemetry*"]

[tool.setuptools.package-data]
wirlwind_telemetry = ["templates/**/*.yaml", "dashboard/*.html"]
